+++ eventio/__init__.py
@@ -1,3 +1,3 @@
-from .event_io_file import objects
-from .iact import IACTFile
-__all__ = ['objects', 'IACTFile']
+from .event_io_file import object_tree
+#from .iact import IACTFile
+__all__ = ['objects',]# 'IACTFile']
